Added Product seeder and "real" data generation
Set up migration auto apply for all environments, as for seeder only for dev env

Added new migrations

Fixed data type for product, price is now decimal instead of double

EF changes, created and updated datetime are automatically audited

+++ src/ZeissAssessment.Application/Contracts/Products/ProductResponse.cs
@@ -5,7 +5,7 @@ public class ProductResponse
     public required int Id { get; init; }
     public required string Name { get; init; }
     public required string Description { get; init; }
-    public required double Price { get; init; }
+    public required decimal Price { get; init; }
     public required int Stock { get; init; }
     public DateTime Created { get; init; }
     public DateTime Updated { get; init; }
