+++ src/ZeissAssessment.Application/Filters/ProductFilters.cs
@@ -3,6 +3,6 @@ namespace ZeissAssessment.Application.Filters;
 public class ProductSearchFilter
 {
     public string? Name { get; set; }
-    public double? MaxPrice { get; set; }
-    public double? MinPrice { get; set; }
+    public decimal? MaxPrice { get; set; }
+    public decimal? MinPrice { get; set; }
 }
